+++ tests/UI/campaigns/functional/BO/03_catalog/01_products/02_paginationAndSortProducts.js
@@ -14,7 +14,7 @@ const loginCommon = require('@commonTests/loginBO');
 // Import pages
 const dashboardPage = require('@pages/BO/dashboard');
 const productsPage = require('@pages/BO/catalog/products/index');
-const addProductPage = require('@pages/BO/catalog/products/add');
+const addProductPage = require('@pages/BO/catalog/products/add/addProduct');
 
 // Import data
 const ProductFaker = require('@data/faker/product');
