+++ tests/UI/campaigns/functional/BO/03_catalog/01_products/06_bulkActions.js
@@ -16,7 +16,7 @@ const ProductFaker = require('@data/faker/product');
 // Import pages
 const dashboardPage = require('@pages/BO/dashboard');
 const productsPage = require('@pages/BO/catalog/products');
-const addProductPage = require('@pages/BO/catalog/products/add');
+const addProductPage = require('@pages/BO/catalog/products/add/addProduct');
 
 const baseContext = 'functional_BO_catalog_products_bulkActions';
