+++ tests/UI/campaigns/functional/BO/03_catalog/03_monitoring/sortAndPagination/04_disabledProducts.js
@@ -14,7 +14,7 @@ const loginCommon = require('@commonTests/loginBO');
 // Import pages
 const dashboardPage = require('@pages/BO/dashboard');
 const productsPage = require('@pages/BO/catalog/products');
-const addProductPage = require('@pages/BO/catalog/products/add');
+const addProductPage = require('@pages/BO/catalog/products/add/addProduct');
 const monitoringPage = require('@pages/BO/catalog/monitoring');
 
 const ProductFaker = require('@data/faker/product');
