+++ tests/UI/campaigns/functional/BO/13_shopParameters/03_productSettings/03_productsStock/08_defaultPackStockManagement.js
@@ -13,7 +13,7 @@ const loginCommon = require('@commonTests/loginBO');
 const dashboardPage = require('@pages/BO/dashboard');
 const productSettingsPage = require('@pages/BO/shopParameters/productSettings');
 const productsPage = require('@pages/BO/catalog/products');
-const addProductPage = require('@pages/BO/catalog/products/add');
+const addProductPage = require('@pages/BO/catalog/products/add/addProduct');
 
 // Import FO pages
 const foProductPage = require('@pages/FO/product');
