+++ tests/UI/campaigns/functional/BO/15_header/02_quickAccess.js
@@ -11,7 +11,7 @@ const dashboardPage = require('@pages/BO/dashboard');
 const statsPage = require('@pages/BO/stats');
 const moduleManagerPage = require('@pages/BO/modules/moduleManager');
 const newCategoryPage = require('@pages/BO/catalog/categories/add');
-const newProductPage = require('@pages/BO/catalog/products/add');
+const addProductPage = require('@pages/BO/catalog/products/add/addProduct');
 const newVoucherPage = require('@pages/BO/catalog/discounts/add');
 const ordersPage = require('@pages/BO/orders');
 const quickAccessPage = require('@pages/BO/quickAccess');
