+++ tests/UI/campaigns/sanity/03_productsBO/03_CRUDStandardProductWithCombinationsInBO.js
@@ -12,7 +12,7 @@ const loginCommon = require('@commonTests/loginBO');
 // Import BO pages
 const dashboardPage = require('@pages/BO/dashboard');
 const productsPage = require('@pages/BO/catalog/products');
-const addProductPage = require('@pages/BO/catalog/products/add');
+const addProductPage = require('@pages/BO/catalog/products/add/addProduct');
 
 // Import FO pages
 const foProductPage = require('@pages/FO/product');
