+++ tests/UI/campaigns/sanity/03_productsBO/04_deleteProduct.js
@@ -12,7 +12,7 @@ const loginCommon = require('@commonTests/loginBO');
 // Import pages
 const dashboardPage = require('@pages/BO/dashboard');
 const productsPage = require('@pages/BO/catalog/products');
-const addProductPage = require('@pages/BO/catalog/products/add');
+const addProductPage = require('@pages/BO/catalog/products/add/addProduct');
 
 // Import data
 const ProductFaker = require('@data/faker/product');
